# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TradeXV2 API",
    "description": "Quantitative Trading Platform API - Market Data, Analytics, Scanner, Strategy, Replay, Backtesting",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/scanner/results": {
      "get": {
        "tags": [
          "Scanner"
        ],
        "summary": "Get Scan Results",
        "description": "Get historical scanner results from scan store.\n\nReturns completed scan results with candidates and metrics.\nUses real DuckDB scan_store for persistence.",
        "operationId": "get_scan_results_api_v1_scanner_results_get",
        "parameters": [
          {
            "name": "scanner_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by scanner name",
              "title": "Scanner Name"
            },
            "description": "Filter by scanner name"
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Date in YYYY-MM-DD format",
              "title": "Date"
            },
            "description": "Date in YYYY-MM-DD format"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max results",
              "default": 10,
              "title": "Limit"
            },
            "description": "Max results"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Scan Results Api V1 Scanner Results Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scanner/top-candidates": {
      "get": {
        "tags": [
          "Scanner"
        ],
        "summary": "Get Top Candidates",
        "description": "Get top scanner candidates.\n\nDelegates to analytics endpoint for v_top3/v_top10.",
        "operationId": "get_top_candidates_api_v1_scanner_top_candidates_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Max candidates",
              "default": 10,
              "title": "Limit"
            },
            "description": "Max candidates"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScannerCandidatesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scanner/snapshots": {
      "get": {
        "tags": [
          "Scanner"
        ],
        "summary": "Get Snapshots",
        "description": "Get full intraday scanner snapshots.\n\nReturns v_intraday_snapshot for all active symbols.",
        "operationId": "get_snapshots_api_v1_scanner_snapshots_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Max symbols",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max symbols"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScannerCandidatesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scanner/run": {
      "post": {
        "tags": [
          "Scanner"
        ],
        "summary": "Run Scan",
        "description": "Trigger a new scanner run.\n\nExecutes ScannerRunner.run_all() for the specified scanner\nand universe. Returns scan ID for result retrieval.\nUses real ScannerRunner with scanner implementations.",
        "operationId": "run_scan_api_v1_scanner_run_post",
        "parameters": [
          {
            "name": "scanner_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scanner to run",
              "title": "Scanner Name"
            },
            "description": "Scanner to run"
          },
          {
            "name": "universe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Universe to scan",
              "default": "NIFTY500",
              "title": "Universe"
            },
            "description": "Universe to scan"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Scan Api V1 Scanner Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScannerCandidatesResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ScannerSnapshot"
            },
            "type": "array",
            "title": "Candidates"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "candidates",
          "count"
        ],
        "title": "ScannerCandidatesResponse",
        "description": "Top scanner candidates."
      },
      "ScannerSnapshot": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "ltp": {
            "type": "number",
            "title": "Ltp"
          },
          "intraday_score": {
            "type": "number",
            "title": "Intraday Score"
          },
          "signal": {
            "type": "string",
            "title": "Signal"
          },
          "trend": {
            "type": "string",
            "title": "Trend"
          },
          "momentum_5d_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Momentum 5D Pct",
            "description": "5-day price change %; not RSI."
          },
          "rsi_14": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rsi 14",
            "description": "True RSI(14) when sourced from feature views; absent on intraday snapshot."
          },
          "roc_5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roc 5"
          },
          "relative_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Volume"
          },
          "day_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day High"
          },
          "day_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Low"
          },
          "day_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Volume"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "ltp",
          "intraday_score",
          "signal",
          "trend"
        ],
        "title": "ScannerSnapshot",
        "description": "Intraday scanner snapshot for a symbol."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}